requirements.txt
pdfplumber>=0.7.5
pdf2image>=1.16.0
pytesseract>=0.3.8
Pillow>=8.0.0
tqdm>=4.60.0
PyMuPDF>=1.18.19
